% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{GMM}
\alias{GMM}
\title{EM Gaussian Mixture Model}
\usage{
GMM(data.mat, K)
}
\arguments{
\item{data.mat}{numeric input data matrix [1 x n]}

\item{K}{int desired number of clusters}
}
\value{
Output list including the log likelihood and the cluster assignment vector [1 x n]
}
\description{
EM Gaussian Mixture Model
}
\examples{
data.mat <- as.matrix(iris[, c("Petal.Length", "Petal.Width")])
K <- 3
GMM.results <- RPackage1::GMM(data.mat, K)
}
